+++ function-aws-custom-runtime/src/test/groovy/io/micronaut/function/aws/runtime/RuntimeApiSpec.groovy
@@ -1,31 +1,13 @@
-/*
- * Copyright 2017-2019 original authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 package io.micronaut.function.aws.runtime
 
 import com.amazonaws.serverless.proxy.model.ApiGatewayRequestIdentity
 import com.amazonaws.serverless.proxy.model.AwsProxyRequest
 import com.amazonaws.serverless.proxy.model.AwsProxyRequestContext
-import com.amazonaws.serverless.proxy.model.AwsProxyResponse
 import com.amazonaws.services.lambda.runtime.Context
 import io.micronaut.context.ApplicationContext
 import io.micronaut.context.BeanProvider
 import io.micronaut.context.annotation.Any
 import io.micronaut.context.annotation.Requires
-import io.micronaut.function.aws.MicronautRequestHandler
-import io.micronaut.http.HttpHeaders
 import io.micronaut.http.HttpResponse
 import io.micronaut.http.MediaType
 import io.micronaut.http.annotation.Body
@@ -37,42 +19,15 @@ import io.micronaut.http.annotation.Produces
 import io.micronaut.runtime.server.EmbeddedServer
 import spock.lang.Specification
 import spock.util.concurrent.PollingConditions
-import spock.util.environment.RestoreSystemProperties
 
-class MicronautLambdaRuntimeSpec extends Specification {
-
-    @RestoreSystemProperties
-    void "Tracing header propagated as system property"() {
+class RuntimeApiSpec extends Specification {
+    void "test runtime API loop"() {
         given:
-        String traceHeader = 'Root=1-5759e988-bd862e3fe1be46a994272793;Sampled=1'
-        EmbeddedServer embeddedServer = ApplicationContext.run(EmbeddedServer, ['spec.name': 'MicronautLambdaRuntimeSpec'])
+        EmbeddedServer embeddedServer = ApplicationContext.run(EmbeddedServer, ['spec.name': 'RuntimeApiSpec'])
         String serverUrl = "localhost:$embeddedServer.port"
         CustomMicronautLambdaRuntime customMicronautLambdaRuntime = new CustomMicronautLambdaRuntime(serverUrl)
         Thread t = new Thread({ ->
             customMicronautLambdaRuntime.run([] as String[])
-        })
-        t.start()
-
-        when:
-        def httpHeaders = Stub(HttpHeaders) {
-            get(LambdaRuntimeInvocationResponseHeaders.LAMBDA_RUNTIME_TRACE_ID) >> traceHeader
-        }
-        customMicronautLambdaRuntime.propagateTraceId(httpHeaders)
-
-        then:
-        System.getProperty(MicronautRequestHandler.LAMBDA_TRACE_HEADER_PROP) == traceHeader
-
-        cleanup:
-        embeddedServer.close()
-    }
-
-    void "test runtime API loop"() {
-        given:
-        EmbeddedServer embeddedServer = ApplicationContext.run(EmbeddedServer, ['spec.name': 'MicronautLambdaRuntimeSpec'])
-        String serverUrl = "localhost:$embeddedServer.port"
-        CustomMicronautLambdaRuntime customMicronautLambdaRuntime = new CustomMicronautLambdaRuntime(serverUrl)
-        Thread t = new Thread({ ->
-             customMicronautLambdaRuntime.run([] as String[])
         })
         t.start()
 
@@ -117,7 +72,7 @@ class MicronautLambdaRuntimeSpec extends Specification {
         }
     }
 
-    @Requires(property = 'spec.name', value = 'MicronautLambdaRuntimeSpec')
+    @Requires(property = 'spec.name', value = 'RuntimeApiSpec')
     @Controller("/")
     static class MockLambadaRuntimeApi {
 
@@ -133,7 +88,7 @@ class MicronautLambdaRuntimeSpec extends Specification {
             context.setIdentity(new ApiGatewayRequestIdentity())
             req.setRequestContext(context)
             HttpResponse.ok(req)
-                .header(LambdaRuntimeInvocationResponseHeaders.LAMBDA_RUNTIME_AWS_REQUEST_ID, "123456")
+                    .header(LambdaRuntimeInvocationResponseHeaders.LAMBDA_RUNTIME_AWS_REQUEST_ID, "123456")
         }
 
         @Post("/2018-06-01/runtime/invocation/{requestId}/response")
@@ -142,5 +97,4 @@ class MicronautLambdaRuntimeSpec extends Specification {
             return HttpResponse.accepted()
         }
     }
-
 }
